% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{covboost-package}
\alias{covboost-package}
\title{Estimate Covariance Matrices Through Boosting}
\description{
covboost
}
\details{
\code{covboost} estimates covariance matrices through boosting.
This allows automatic sparsity detection. The biggest benefits are perhaps
for high dimensional \code{p>>n} problems.

Important functions:

\itemize{
\item \code{\link{covboost}}: function for boosting out a covariance matrix from the Identity matrix
\item \code{\link{covboost_cv}}: The same as \code{\link{covboost}}, but adds k-fold cross validation
}

See individual function documentation for usage.
}
\author{
Berent Ånund Strømnes Lunde
}
