% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covboost.R
\name{covboost}
\alias{covboost}
\title{Add together two numbers}
\usage{
covboost(x, learning_rate = 0.01, niter = 1000)
}
\arguments{
\item{x}{An \code{n X p} matrix or data frame of dimensions n}

\item{learning_rate}{Scaling the path of elements in the covariance matrix}

\item{niter}{The number of boosting iterations}
}
\value{
A list of useful boosting information, most importantly the covariance matrix
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
