% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covboost_cv.R
\name{covboost_cv}
\alias{covboost_cv}
\title{Boosted Covariance Matrix Estimation With Boosting}
\usage{
covboost_cv(x, learning_rate = 0.01, niter = 1000, nfolds = 10)
}
\arguments{
\item{x}{An \code{n x p} numeric matrix or data frame}

\item{learning_rate}{Scaling the path of elements in the covariance matrix}

\item{niter}{The number of boosting iterations}

\item{nfolds}{The number of folds for k-fold cross validation}
}
\value{
A list of useful cross validation information from boosting out the covariance matrix
}
\description{
Boosted Covariance Matrix Estimation With Boosting
}
\examples{
\dontrun{
## Generate data with random correlation
p <- 50
n <- 50
x <- matrix(nrow=n, ncol=p)
x[,1] <- rnorm(n)
for(j in 2:p){
    rho <- runif(1,-1,1)*rbinom(1,1,0.5)
    col_dependence <- sample(j-1,1)
    x[,j] <- rnorm(n, mean=rho*x[,col_dependence], sd=(1-rho^2))
}
lrn_rate <- 0.2
cov_cv <- covboost_cv(x, learning_rate=lrn_rate)
con_cv$cvplot
con_cv$opt_iter
}

}
