% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covboost.R
\name{covboost}
\alias{covboost}
\title{Boosted Covariance Matrix Estimation}
\usage{
covboost(x, shrinkage = 0.1, learning_rate = 0.5, niter = 1000)
}
\arguments{
\item{x}{An \code{n x p} numeric matrix or data frame}

\item{shrinkage}{Shrinkage defining the trade-off between the Identity and variance matrices}

\item{learning_rate}{Scaling the path of elements in the covariance matrix}

\item{niter}{The number of boosting iterations}
}
\value{
A list of useful boosting information, most importantly the covariance matrix
}
\description{
Boosted Covariance Matrix Estimation
}
\examples{
\dontrun{
## Generate data with random correlation and estimate
p <- 50
n <- 50
x <- matrix(nrow=n, ncol=p)
x[,1] <- rnorm(n)
for(j in 2:p){
    rho <- runif(1,-1,1)*rbinom(1,1,0.5)
    col_dependence <- sample(j-1,1)
    x[,j] <- rnorm(n, mean=rho*x[,col_dependence], sd=(1-rho^2))
}
lrn_rate <- 0.2
cov_cv <- covboost_cv(x, learning_rate=lrn_rate)
cov_cv$cvplot
cov_cv$opt_iter
cov_cv$opt_shrinkage
sigma <- covboost(x, shrinkage=cov_cv$opt_shrinkage, niter=cov_cv$opt_iter, learning_rate=lrn_rate)
sigma$cov
sigma$plot
}
}
