% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covboost-package.R, R/covboost.R
\docType{package}
\name{covboost}
\alias{covboost}
\alias{covboost-package}
\title{Estimate Covariance Matrices Through Boosting}
\usage{
covboost(x, learning_rate = 0.01, niter = 1000)
}
\arguments{
\item{x}{An \code{n x p} numeric matrix or data frame}

\item{learning_rate}{Scaling the path of elements in the covariance matrix}

\item{niter}{The number of boosting iterations}
}
\value{
A list of useful boosting information, most importantly the covariance matrix
}
\description{
Estimate Covariance Matrices Through Boosting
}
\details{
\code{covboost} estimates covariance matrices through boosting.
This allows automatic sparsity detection. The biggest benefits are perhaps
for high dimensional \code{p>>n} problems.

Important functions:

\itemize{
\item \code{\link{covboost}}: function for boosting out a covariance matrix from the Identity matrix
\item \code{\link{covboost_cv}}: The same as \code{\link{covboost}}, but adds k-fold cross validation
}

See individual function documentation for usage.
}
\examples{
\dontrun{
## Generate data with random correlation and estimate
p <- 50
n <- 50
x <- matrix(nrow=n, ncol=p)
x[,1] <- rnorm(n)
for(j in 2:p){
    rho <- runif(1,-1,1)*rbinom(1,1,0.5)
    col_dependence <- sample(j-1,1)
    x[,j] <- rnorm(n, mean=rho*x[,col_dependence], sd=(1-rho^2))
}
lrn_rate <- 0.2
cov_cv <- covboost_cv(x, learning_rate=lrn_rate)
cov_cv$cvplot
cov_cv$opt_iter
sigma <- covboost(x, niter=cov_cv$opt_iter, learning_rate=lrn_rate)
sigma$cov
sigma$plot
}
}
\author{
Berent Ånund Strømnes Lunde
}
