% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covboost-package.R, R/covboost.R
\docType{package}
\name{covboost}
\alias{covboost}
\alias{covboost-package}
\title{Estimate Covariance Matrices Through Boosting}
\usage{
covboost(x, learning_rate = 0.01, niter = 1000)
}
\arguments{
\item{x}{An \code{n X p} matrix or data frame of dimensions n}

\item{learning_rate}{Scaling the path of elements in the covariance matrix}

\item{niter}{The number of boosting iterations}
}
\value{
A list of useful boosting information, most importantly the covariance matrix
}
\description{
Estimate Covariance Matrices Through Boosting
}
\details{
\code{covboost} estimates covariance matrices through boosting.
This allows automatic sparsity detection. The biggest benefits are perhaps
for high dimensional \code{p>>n} problems.

Important functions:

\itemize{
\item \code{\link{covboost}}: function for boosting out a covariance matrix from the Identity matrix
\item \code{\link{covboost_cv}}: The same as \code{\link{covboost}}, but adds k-fold cross validation
}

See individual function documentation for usage.
}
\examples{
add(1, 1)
add(10, 1)
}
\author{
Berent Ånund Strømnes Lunde
}
