% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covboost_cv.R
\name{covboost_cv}
\alias{covboost_cv}
\title{Boosted Covariance Matrix Estimation With CV}
\usage{
covboost_cv(x, learning_rate = 0.5, niter = 1000, nfolds = 10,
  cores = 1)
}
\arguments{
\item{x}{An \code{n x p} numeric matrix or data frame}

\item{learning_rate}{Scaling the path of elements in the covariance matrix}

\item{niter}{The number of boosting iterations}

\item{nfolds}{The number of folds for k-fold cross validation}

\item{cores}{The number of cores use for parallel computations}
}
\value{
A list of useful cross validation information from boosting out the covariance matrix
}
\description{
Boosted Covariance Matrix Estimation With CV
}
\examples{
\dontrun{
## Generate data with random correlation
p <- 50
n <- 50
x <- matrix(nrow=n, ncol=p)
x[,1] <- rnorm(n)
for(j in 2:p){
    rho <- runif(1,-1,1)*rbinom(1,1,0.5)
    col_dependence <- sample(j-1,1)
    x[,j] <- rnorm(n, mean=rho*x[,col_dependence], sd=(1-rho^2))
}
lrn_rate <- 0.2
cov_cv <- covboost_cv(x, learning_rate=lrn_rate)
cov_cv$cvplot
cov_cv$opt_iter
}

}
