% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covboost_cv.R
\name{covboost_cv}
\alias{covboost_cv}
\title{Add together two numbers}
\usage{
covboost_cv(x, learning_rate = 0.01, niter = 1000, nfolds = 10)
}
\arguments{
\item{x}{An \code{n x p} numeric matrix or data frame}

\item{learning_rate}{Scaling the path of elements in the covariance matrix}

\item{niter}{The number of boosting iterations}

\item{nfolds}{The number of folds for k-fold cross validation}
}
\value{
A list of useful cross validation information from boosting out the covariance matrix
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
